server/requirement.txt
python-dotenv
mysql-connector-python
Flask